#57 updated files structure
